src/requirements.txt
rembg==2.0.69
onnxruntime==1.23.2
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# Улучшение лица
gfpgan==1.3.8
torch==2.9.1
torchvision==0.24.1
basicsr==1.4.2
facexlib==0.3.0

# Нужные зависимые пакеты (в версиях из текущего окружения)
scipy==1.15.3
scikit-image==0.25.2
tqdm==4.67.1
requests==2.32.5
matplotlib==3.10.8

